+++ andes/models/exciter/esst4b.py
@@ -198,6 +198,7 @@ class ESST4BModel(ExcBase):
 
         self.VB = GainLimiter(u='VE*FEX_y',
                               K=1,
+                              R=1,
                               upper=self.VBMAX,
                               lower=self.VBMIN,
                               no_lower=True,
@@ -206,6 +207,7 @@ class ESST4BModel(ExcBase):
 
         self.VG = GainLimiter(u=self.vout,
                               K=self.KG,
+                              R=1,
                               upper=self.VGMAX,
                               lower=self.VGMIN,
                               no_lower=True,